+++ event/event-service-with-loco/benches/matching_bench.rs
@@ -74,8 +74,8 @@ fn bench_location_matching(c: &mut Criterion) {
             postal_code: Some("94720".into()),
             country: Some("US".into()),
         }),
-        latitude: Some(37.873),
-        longitude: Some(-122.254),
+        latitude: Some("37.873".parse().unwrap()),
+        longitude: Some("-122.254".parse().unwrap()),
         url: None,
     };
     let p2 = p1.clone();